// src/fine/interfaces/fine_details.interface.ts
export interface IFineDetails {
    id: string;
    plateNumber: string;
    evidenceCID: string;
    location: string;
    timestamp: string;
    infractionType: string;
    cost: number;
    ownerIdentifier: string;
    currentState: string;
    registeredBy: string;
    externalSystemId: string;
    hashImageIPFS: string;
}